+++ Soulgram.Interests/Soulgram.Interests.Benchmark/ServiceLocator.cs
@@ -4,5 +4,5 @@ namespace Soulgram.Interests.Api.Benchmark;
 
 public static class ServiceLocator
 {
-    public static ServiceProvider ServiceProvider { get; set; }
+    public static ServiceProvider ServiceProvider { get; set; } = null!;
 }
