+++ Soulgram.Interests/Soulgram.Interests.Infrastructure/Models/MainClientResponses/Caption.cs
@@ -5,8 +5,8 @@ namespace Soulgram.Interests.Infrastructure.Models.MainClientResponses;
 public class Caption
 {
     [JsonPropertyName("plainText")]
-    public string PlainText { get; set; }
+    public string? PlainText { get; set; }
 
     [JsonPropertyName("__typename")]
-    public string Typename { get; set; }
+    public string? Typename { get; set; }
 }
