+++ Soulgram.Interests/Soulgram.Interests.Infrastructure/Models/MainClientResponses/TitleText.cs
@@ -5,8 +5,8 @@ namespace Soulgram.Interests.Infrastructure.Models.MainClientResponses;
 public class TitleText
 {
     [JsonPropertyName("text")]
-    public string Text { get; set; }
+    public string? Text { get; set; }
 
     [JsonPropertyName("__typename")]
-    public string Typename { get; set; }
+    public string? Typename { get; set; }
 }
