+++ Soulgram.Interests/Soulgram.Interests.Infrastructure/Models/MainClientResponses/TitleType.cs
@@ -5,10 +5,10 @@ namespace Soulgram.Interests.Infrastructure.Models.MainClientResponses;
 public class TitleType
 {
     [JsonPropertyName("text")]
-    public string Text { get; set; }
+    public string? Text { get; set; }
 
     [JsonPropertyName("id")]
-    public string Id { get; set; }
+    public string? Id { get; set; }
 
     [JsonPropertyName("isSeries")]
     public bool? IsSeries { get; set; }
@@ -17,5 +17,5 @@ public class TitleType
     public bool? IsEpisode { get; set; }
 
     [JsonPropertyName("__typename")]
-    public string Typename { get; set; }
+    public string? Typename { get; set; }
 }
