+++ Soulgram.Interests/Soulgram.Interests.sln
@@ -12,8 +12,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Soulgram.Interests.Persiste
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{2A63177E-6E1F-4539-A5C2-69E9857158BC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Soulgram.Interests.Benchmark", "Soulgram.Interests.Benchmark\Soulgram.Interests.Benchmark.csproj", "{248DB909-F0F5-4879-B90C-21E73DC20D35}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,12 +38,7 @@ Global
 		{FDA5AC2C-F00C-4076-9863-C126F99D4A05}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FDA5AC2C-F00C-4076-9863-C126F99D4A05}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FDA5AC2C-F00C-4076-9863-C126F99D4A05}.Release|Any CPU.Build.0 = Release|Any CPU
-		{248DB909-F0F5-4879-B90C-21E73DC20D35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{248DB909-F0F5-4879-B90C-21E73DC20D35}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{248DB909-F0F5-4879-B90C-21E73DC20D35}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{248DB909-F0F5-4879-B90C-21E73DC20D35}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{248DB909-F0F5-4879-B90C-21E73DC20D35} = {2A63177E-6E1F-4539-A5C2-69E9857158BC}
 	EndGlobalSection
 EndGlobal
